# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restomate API",
        "description": "API de réservation de tables de restaurant. Gérez restaurants, tables et réservations avec authentification Sanctum.",
        "contact": {
            "name": "Alex AGO",
            "email": "alexyessougnonago@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://restomate-api.test/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Toutes les réservations (Admin)",
                "operationId": "c0ead7ee8da750a341eeff509f3af8b9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "cancelled",
                                "no_show"
                            ]
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-01"
                        }
                    },
                    {
                        "name": "restaurant_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservations récupérées"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Créer une réservation (Client)",
                "operationId": "ab4d7952bb3ef1ccd7724e8d862b034b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_id",
                                    "reservation_date",
                                    "start_time",
                                    "number_of_guests"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-01"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "19:30"
                                    },
                                    "number_of_guests": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Allergie aux noix, anniversaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée"
                    },
                    "409": {
                        "description": "Table non disponible"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/my": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Mes réservations (Client connecté)",
                "operationId": "da10b02490ecac708ba3fcd0f2de94a6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "cancelled",
                                "no_show"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservations récupérées"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Détail d'une réservation",
                "operationId": "49499ada3f4d3680e147782d33573252",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation récupérée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Modifier une réservation",
                "operationId": "4a219e5cfd9760d12da084929201d042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-02"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "20:00"
                                    },
                                    "number_of_guests": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Table près de la fenêtre"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation mise à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Impossible de modifier"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}/cancel": {
            "patch": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Annuler une réservation",
                "operationId": "70237665fb6d7ef14914c29ac4190d5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation annulée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Annulation impossible"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}/confirm": {
            "patch": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Confirmer une réservation (Admin)",
                "operationId": "adeb83734efe5b5a98da3b1bd237a791",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation confirmée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Statut incorrect"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}/no-show": {
            "patch": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Marquer no-show (Admin)",
                "description": "Marque une réservation comme no-show quand le client ne s'est pas présenté",
                "operationId": "8ccacf5e373317293cff33e8cdccf3cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marqué no-show"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Statut incorrect"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entre ton token Sanctum obtenu après login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}